Add files via upload
This program calculates and displays the simple interest, compound interest, and maturity value of a mutual fund based on the principal amount, rate of return, and time in years. You can run this program, enter the required details, and it will display the calculated values. Note: The rate of return should be entered as a percentage, e.g., for 10% enter 10.